+++ Pdf4QtEditor/main.cpp
@@ -101,7 +101,7 @@ QCoreApplication::setApplicationName("FamilyPDF Editor");
 
     pdf::PDFWidgetUtils::setDarkTheme(isLightGui, isDarkGui);
 
-    QIcon appIcon(":/app-icon.svg");
+    QIcon appIcon(":/familypdf-app-icon.svg");
     QApplication::setWindowIcon(appIcon);
 
     const QStringList savedSession = pdfviewer::PDFSessionManager::loadPaths();